requirements.txt
fastapi
uvicorn
aiogram
sqlalchemy
aiosqlite
python-dotenv
pydantic-settings
requests
httpx
